Added optimize function to api, cleaned up the code
